reorganize event handler code
