+++ client/secrets.go
@@ -8,9 +8,45 @@ import (
 	log "github.com/sirupsen/logrus"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/watch"
 
 	"k8s.io/client-go/kubernetes"
 )
+
+func secretEventHandler(ctx context.Context, clientset *kubernetes.Clientset, config *SyncConfig, event watch.Event) {
+	secret := event.Object.(*v1.Secret)
+
+	switch event.Type {
+	case watch.Added:
+		addSecrets(ctx, clientset, config, secret)
+	case watch.Modified:
+		modifySecrets(ctx, clientset, config, secret)
+	case watch.Deleted:
+		deleteSecrets(ctx, clientset, config, secret)
+	}
+}
+
+func addSecrets(ctx context.Context, clientset *kubernetes.Clientset, config *SyncConfig, secret *v1.Secret) error {
+	log.Infof("[%s/%s]: Secret created", secret.ObjectMeta.Namespace, secret.ObjectMeta.Name)
+
+	return syncNamespaceSecret(ctx, clientset, config, secret, syncAddedModifiedSecret)
+}
+
+func modifySecrets(ctx context.Context, clientset *kubernetes.Clientset, config *SyncConfig, secret *v1.Secret) error {
+	if secret.DeletionTimestamp != nil {
+		return nil
+	}
+
+	log.Infof("[%s/%s]: Secret modified", secret.ObjectMeta.Namespace, secret.ObjectMeta.Name)
+
+	return syncNamespaceSecret(ctx, clientset, config, secret, syncAddedModifiedSecret)
+}
+
+func deleteSecrets(ctx context.Context, clientset *kubernetes.Clientset, config *SyncConfig, secret *v1.Secret) error {
+	log.Infof("[%s/%s]: Secret deleted", secret.ObjectMeta.Namespace, secret.ObjectMeta.Name)
+
+	return syncNamespaceSecret(ctx, clientset, config, secret, syncDeletedSecret)
+}
 
 type SecretSyncFunc func(context.Context, *kubernetes.Clientset, *SyncConfig, v1.Namespace, *v1.Secret) error
