+++ client/sync.go
@@ -4,10 +4,7 @@ import (
 	"context"
 
 	log "github.com/sirupsen/logrus"
-	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/watch"
-	"k8s.io/client-go/kubernetes"
 )
 
 // SyncSecrets syncs Secrets across all selected Namespaces
@@ -39,52 +36,4 @@ func SyncSecrets(config *SyncConfig) (err error) {
 			namespaceEventHandler(ctx, clientset, config, namespaceEvent)
 		}
 	}
-}
-
-func secretEventHandler(ctx context.Context, clientset *kubernetes.Clientset, config *SyncConfig, event watch.Event) {
-	secret := event.Object.(*v1.Secret)
-
-	switch event.Type {
-	case watch.Added:
-		addSecrets(ctx, clientset, config, secret)
-	case watch.Modified:
-		modifySecrets(ctx, clientset, config, secret)
-	case watch.Deleted:
-		deleteSecrets(ctx, clientset, config, secret)
-	}
-}
-
-func addSecrets(ctx context.Context, clientset *kubernetes.Clientset, config *SyncConfig, secret *v1.Secret) error {
-	log.Infof("[%s/%s]: Secret created", secret.ObjectMeta.Namespace, secret.ObjectMeta.Name)
-
-	return syncNamespaceSecret(ctx, clientset, config, secret, syncAddedModifiedSecret)
-}
-
-func modifySecrets(ctx context.Context, clientset *kubernetes.Clientset, config *SyncConfig, secret *v1.Secret) error {
-	if secret.DeletionTimestamp != nil {
-		return nil
-	}
-
-	log.Infof("[%s/%s]: Secret modified", secret.ObjectMeta.Namespace, secret.ObjectMeta.Name)
-
-	return syncNamespaceSecret(ctx, clientset, config, secret, syncAddedModifiedSecret)
-}
-
-func deleteSecrets(ctx context.Context, clientset *kubernetes.Clientset, config *SyncConfig, secret *v1.Secret) error {
-	log.Infof("[%s/%s]: Secret deleted", secret.ObjectMeta.Namespace, secret.ObjectMeta.Name)
-
-	return syncNamespaceSecret(ctx, clientset, config, secret, syncDeletedSecret)
-}
-
-func namespaceEventHandler(ctx context.Context, clientset *kubernetes.Clientset, config *SyncConfig, event watch.Event) {
-	namespace := event.Object.(*v1.Namespace)
-
-	switch event.Type {
-	case watch.Added:
-		addNamespace(ctx, clientset, config, namespace)
-	}
-}
-
-func addNamespace(ctx context.Context, clientset *kubernetes.Clientset, config *SyncConfig, namespace *v1.Namespace) {
-	log.Infof("[%s]: Namespace created", namespace.Name)
 }
